Rename token pages to {name}TokenPage
