--- src/pages/tokenLending/BorrowPage/BorrowPage.tsx
@@ -1,3 +1,0 @@
-export const BorrowPage = () => {
-  return <div>BorrowPage</div>
-}
