--- src/pages/tokenLending/BorrowPage/index.ts
@@ -1,1 +1,0 @@
-export { BorrowPage } from './BorrowPage'
