+++ src/pages/tokenLending/BorrowTokenPage/BorrowTokenPage.tsx
@@ -1,0 +1,3 @@
+export const BorrowTokenPage = () => {
+  return <div>BorrowTokenPage</div>
+}
