+++ src/pages/tokenLending/BorrowTokenPage/index.ts
@@ -1,0 +1,1 @@
+export { BorrowTokenPage } from './BorrowTokenPage'
