--- src/pages/tokenLending/LendPage/index.ts
@@ -1,1 +1,0 @@
-export { LendPage } from './LendPage'
