+++ src/pages/tokenLending/LendTokenPage/index.ts
@@ -1,0 +1,1 @@
+export { LendTokenPage } from './LendTokenPage'
