--- src/pages/tokenLending/LoansPage/LoansPage.tsx
@@ -1,3 +1,0 @@
-export const LoansPage = () => {
-  return <div>LoansPage</div>
-}
