--- src/pages/tokenLending/LoansPage/index.ts
@@ -1,1 +1,0 @@
-export { LoansPage } from './LoansPage'
