+++ src/pages/tokenLending/LoansTokenPage/LoansTokenPage.tsx
@@ -1,0 +1,3 @@
+export const LoansTokenPage = () => {
+  return <div>LoansTokenPage</div>
+}
