+++ src/pages/tokenLending/LoansTokenPage/index.ts
@@ -1,0 +1,1 @@
+export { LoansTokenPage } from './LoansTokenPage'
