--- src/pages/tokenLending/OffersPage/OffersPage.tsx
@@ -1,3 +1,0 @@
-export const OffersPage = () => {
-  return <div>OffersPage</div>
-}
