--- src/pages/tokenLending/OffersPage/index.ts
@@ -1,1 +1,0 @@
-export { OffersPage } from './OffersPage'
