+++ src/pages/tokenLending/OffersTokenPage/OffersTokenPage.tsx
@@ -1,0 +1,3 @@
+export const OffersTokenPage = () => {
+  return <div>OffersTokenPage</div>
+}
