+++ src/pages/tokenLending/OffersTokenPage/index.ts
@@ -1,0 +1,1 @@
+export { OffersTokenPage } from './OffersTokenPage'
