+++ src/pages/tokenLending/index.ts
@@ -1,4 +1,4 @@
-export { BorrowPage as BorrowTokenPage } from './BorrowPage'
-export { LendPage as LendTokenPage } from './LendPage'
-export { LoansPage as LoansTokenPage } from './LoansPage'
-export { OffersPage as OffersTokenPage } from './OffersPage'
+export { BorrowTokenPage } from './BorrowTokenPage'
+export { LendTokenPage } from './LendTokenPage'
+export { LoansTokenPage } from './LoansTokenPage'
+export { OffersTokenPage } from './OffersTokenPage'
